oop/auth_service.py/requirements.txt
passlib[bcrypt]
pyjwt